requirements.txt
pandas 
openpyxl 
google-generativeai
python-dotenv
streamlit
pymongo